OCLM-47 Introduce import without validation
