+++ api/src/main/java/org/openmrs/module/openconceptlab/ValidationType.java
@@ -1,0 +1,5 @@
+package org.openmrs.module.openconceptlab;
+
+public enum ValidationType {
+    NONE, FULL;
+}
