+++ owa/app/js/subscription/subscription.html
@@ -12,6 +12,7 @@
                     <input id="subscription-token" ng-model="vm.subscription.token" style="width: 60%" required/>
                 </p>
                 <input ng-model="vm.subscription.subscribedToSnapshot" type="checkbox"> {{"subscription.options.checkbox" | translate}}<br>
+                <input ng-model="vm.subscription.validationType" type="checkbox" ng-true-value="'NONE'" ng-false-value="'FULL'"> {{"subscription.validationType.checkbox" | translate}}<br>
             </div>
             <br>
             <div>
