+++ owa/app/translation/messages_en.json
@@ -7,6 +7,7 @@
   "instalupdatesautomatically.radio":"Install updates automatically",
   "subscribe.button":"Save changes",
   "subscription.options.checkbox":"Subscribe to SNAPSHOT versions (not recommended)",
+  "subscription.validationType.checkbox":"Disable validation (should be used with care for well curated collections or sources)",
   "cancelchanges.button":"Cancel Changes",
   "nosubscriptionmessage.info":"You are not subscribed to Open Concept Lab. Please go to the configuration page to setup the subscription",
   "installupdatesperiod.text":"Install updates every",
